Wip: 주문 조회
- 결제 조회
- 주문 조회

+++ src/main/java/com/example/shinsekai/purchase/dto/in/PurchaseRequestDto.java
@@ -23,8 +23,8 @@ public class PurchaseRequestDto {
     private String address;
     private String giftCertificationUuid;
     private String couponUuid;
-    private int shipmentFee;
-    private double productTotalPrice;
+    private Double shipmentFee;
+    private Double productTotalPrice;
 
     public Purchase toEntity(){
         return Purchase.builder()
